+++ lcdspi/lcdspi.h
@@ -125,9 +125,9 @@ extern void lcd_reset_coords();
 //	Customs
 extern void lcd_set_colours(unsigned int fc, unsigned int bc);
 extern void lcd_clear_bottom();
-extern short lcd_get_current_x();
-extern short lcd_get_current_y();
-extern void lcd_set_coords(short x, short y);
+extern int lcd_get_current_x();
+extern int lcd_get_current_y();
+extern void lcd_set_coords(int x, int y);
 
 extern void draw_rect_spi(int x1, int y1, int x2, int y2, int c);
